Replacing deprecated function (#181)
# Pull Request

Fixes
[https://github.com/insightsengineering/biomarker-catalog/issues/65](https://github.com/insightsengineering/biomarker-catalog/issues/65)

The function summarize_vars() has been deprecated and no longer works.
It was used in the R Markdown of this package. A new function
analyze_vars() is a renamed version of summarize_vars(). All occurrences
of the deprecated function in this package have been replaced in this
pull request.